Restructured the core object.
Previously, developers would have to require the module and then instantiate
a new instance. I wanted to remove the need to call `new Thing();` which did
not server any purpose (because there are no options supported at the moment).
Thus, this commit. I also bumped the version to publish it on NPM.